Extract Report subclasses to separate directory
